EDU-1452 Make user role visible
