+++ src/components/admin/user-accounts-tab.less
@@ -58,27 +58,6 @@
   gap: 0.5em;
 }
 
-.UserAccountsTab-roleInfoIcon {
-  cursor: pointer !important;
-  color: @edu-secondary-color-dark !important;
-}
-
-.UserAccountsTab-roleInfoMarkdown {
-  h1 {
-    font-size: @edu-font-size-lg;
-    font-weight: bold;
-    margin: 2em 0 0.5em 0;
-  }
-
-  hr {
-    margin: 2em 0 0.25em 0;
-
-    & + * {
-      font-size: @edu-font-size-sm;
-    }
-  }
-}
-
 .UserAccountsTab-roleSegmentOption {
   display: grid;
   grid-auto-flow: row;
@@ -125,6 +104,7 @@
     .ant-segmented-thumb {
       background-color: @edu-primary-color-light !important;
     }
+
     .ant-segmented-item-selected {
       color: @edu-white !important;
       background-color: @edu-primary-color-light !important;
