+++ src/components/admin/user-accounts-tab.yml
@@ -23,18 +23,6 @@ userAccountsTab:
   activeAccounts:
     en: Active accounts
     de: Aktive Konten
-  role_user:
-    en: User
-    de: Benutzer
-  role_accredited-author:
-    en: Accred. author
-    de: Akkred. Autor
-  role_maintainer:
-    en: Maintainer
-    de: Maintainer
-  role_admin:
-    en: Admin
-    de: Admin
   unconfirmedAccounts:
     en: Unconfirmed accounts
     de: Unbestätigte Konten
@@ -108,7 +96,7 @@ userAccountsTab:
 
       # Maintainer
 
-      Maintainers have extended content management permissions over public content and can curate documents and the media center resources.
+      Maintainers have extended content management permissions over public content and can curate documents and the media center resources.\
       These permissions include reviewing, approving, protecting, archiving or hard-deleting sections of public documents, as well as deleting media center resources.
 
       # Admin
@@ -122,7 +110,6 @@ userAccountsTab:
       # Benutzer
 
       Benutzer können zu den Inhalten der Plattform beitragen (mit Ausnahme von geschützten oder archivierten Dokumenten), aber sie können lediglich ihre eigenen privaten Inhalte löschen.
-      Users can contribute to the platform's content, with the restriction of protected or archived documents. They can only delete own private content.
 
       # Akkreditierter Autor
 
@@ -130,7 +117,7 @@ userAccountsTab:
 
       # Maintainer
 
-      Maintainer haben erweiterte Berechtigungen in der Verwaltung öffentlicher Inhalte und können Dokumente und Ressourcen in der Mediathek pflegen.
+      Maintainer haben erweiterte Berechtigungen in der Verwaltung öffentlicher Inhalte und können Dokumente und Ressourcen in der Mediathek pflegen.\
       Diese Berechtigungen beinhalten das Reviewing, Bestätigen, Schützen, Archivieren oder das unwiderrufliche Löschen von Abschnitten von öffentlichen Dokumenten, sowie das Löschen von Ressourcen in der Mediathek.
 
       # Admin
