+++ src/components/pages/dashboard.less
@@ -36,6 +36,15 @@
   background-color: @edu-white;
 }
 
+.DashboardPage-profileRole {
+  display: grid;
+  gap: 5px;
+  grid-auto-flow: column;
+  justify-content: start;
+  font-size: @edu-font-size-sm;
+  color: @edu-text-color-secondary;
+}
+
 .DashboardPage-profileInfo {
   margin-top: 55px;
   font-size: @edu-font-size-sm;
