+++ src/components/user-role-info.less
@@ -1,0 +1,22 @@
+.UserRoleInfo {}
+
+.UserRoleInfo-infoIcon {
+  cursor: pointer !important;
+  color: @edu-secondary-color-dark !important;
+}
+
+.UserRoleInfo-modalContent {
+  h1 {
+    font-size: @edu-font-size-lg;
+    font-weight: bold;
+    margin: 2em 0 0.5em 0;
+  }
+
+  hr {
+    margin: 2em 0 0.25em 0;
+
+    & + * {
+      font-size: @edu-font-size-sm;
+    }
+  }
+}
