+++ src/components/user-role-info.yml
@@ -1,0 +1,48 @@
+userRoleInfo:
+  modalTitle:
+    en: Roles in {appName}
+    de: Rollen in {appName}
+  modalContentMarkdown:
+    en: |
+      # User
+
+      Users can contribute to the platform's content, with the restriction of protected or archived documents. They can only delete own private content.
+
+      # Accredited Author
+
+      Accredited authors can create protected public documents with exclusive editing permissions.
+
+      # Maintainer
+
+      Maintainers have extended content management permissions over public content and can curate documents and the media center resources.\
+      These permissions include reviewing, approving, protecting, archiving or hard-deleting sections of public documents, as well as deleting media center resources.
+
+      # Administrator
+
+      Administrators can configure global settings, manage user accounts including storage plans, as well as execute technical maintenance tasks.
+
+      ***
+
+      *Roles in {appName} are cumulative, meaning that each consecutive role in this list has all the permissions of the previous role and adds more permissions on top.*
+    de: |
+      # Benutzer
+
+      Benutzer können zu den Inhalten der Plattform beitragen (mit Ausnahme von geschützten oder archivierten Dokumenten), aber sie können lediglich ihre eigenen privaten Inhalte löschen.
+
+      # Akkreditierter Autor
+
+      Akkreditierte Autoren können geschützte öffentliche Dokumente mit exklusiven Bearbeitungsberechtigungen erstellen.
+
+      # Maintainer
+
+      Maintainer haben erweiterte Berechtigungen in der Verwaltung öffentlicher Inhalte und können Dokumente und Ressourcen in der Mediathek pflegen.\
+      Diese Berechtigungen beinhalten das Reviewing, Bestätigen, Schützen, Archivieren oder das unwiderrufliche Löschen von Abschnitten von öffentlichen Dokumenten, sowie das Löschen von Ressourcen in der Mediathek.
+
+      # Administrator
+
+      Administratoren können globale Einstellungen konfigurieren, Benutzerkonten inklusive Speicherpläne verwalten, sowie technische Instandhaltungstasks ausführen.
+
+      ***
+
+      *Rollen in {appName} sind kumulativ, d.h. dass jede aufeinander folgende Rolle in dieser Liste jeweils sowohl alle Berechtigungen der
+      vorhergehenden Rolle als auch darüber hinausgehende eigene Berechtigungen beinhaltet.*
