+++ src/resources/common.yml
@@ -693,3 +693,15 @@ common:
   dashboardTab_settings:
     en: Settings
     de: Einstellungen
+  role_user:
+    en: User
+    de: Benutzer
+  role_accredited-author:
+    en: Accred. author
+    de: Akkred. Autor
+  role_maintainer:
+    en: Maintainer
+    de: Maintainer
+  role_admin:
+    en: Administrator
+    de: Administrator
