+++ src/styles/main.less
@@ -47,6 +47,7 @@
 @import '../components/credits-footer.less';
 @import '../components/pages/register.less';
 @import '../components/markdown-input.less';
+@import '../components/user-role-info.less';
 @import '../components/pages/dashboard.less';
 @import '../components/loading-section.less';
 @import '../components/section-display.less';
